fix: Add a dynamic version attribute for the `setuptools` build-backend (#188)
The template for the setuptools build-backend had an issue where the project version was declared as a dynamic attribute but the associated pyproject.toml file did not contain a [tool.setuptools.dynamic] TOML table. This caused the version to be returned as 0.0.0, instead of the correct version being read from the specified file.